Move Matcher classes into their own sub-directory.
